feat(SG-460): Tasks

+++ database/migrations/20241030115001-task-type.js
@@ -12,7 +12,7 @@ module.exports = {
 			},
 			{
 				id: {
-					type: Sequelize.SMALLINT,
+					type: Sequelize.INTEGER,
 					primaryKey: true,
 					autoIncrement: true,
 				},
